refactor: ReactivePanningScrollview folders, make ReanimatedTabs lazy
